cleanup lib.rs
