+++ src/bin/rustysd.rs
@@ -3,9 +3,9 @@ fn main() {
         .next()
         .expect("could not get executable name from args");
     if exec_name.ends_with("exec_helper") {
-        rustysd::run_exec_helper();
+        rustysd::entrypoints::run_exec_helper();
     } else if exec_name.ends_with("rustysd") {
-        rustysd::run_service_manager();
+        rustysd::entrypoints::run_service_manager();
     } else {
         eprintln!(
             "Can only start as rustysd or exec_helper. Was: {}",
