+++ src/entrypoints/mod.rs
@@ -1,0 +1,5 @@
+mod exec_helper;
+mod service_manager;
+
+pub use exec_helper::{run_exec_helper, ExecHelperConfig};
+pub use service_manager::run_service_manager;
